requirements.txt
streamlit==1.36.0
Pillow==9.5.0